models/real-esrgan/requirements.txt
numpy==1.26.4
opencv-python
Pillow
torch>=1.7
torchvision>=0.8.0
tqdm
huggingface-hub